backend/requirements.txt
# Python: 3.11.x (repo ../.python-version; backend/pyproject.toml requires-python)
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Cache / Pub-Sub
redis==5.1.1

# AI (LLM via Groq — OpenAI-compatible stack removed; use langchain_groq.ChatGroq)
langchain==0.3.0
langchain-groq==0.2.0
langgraph==0.2.20
langgraph-checkpoint-postgres>=2.0.25,<3
psycopg[binary]>=3.2.0

# Git
gitpython==3.1.43

# HTTP client (for LangChain internals and tests)
httpx==0.27.2

whatthepatch==1.0.6  # D-04: parse unified diff → Monaco original/modified strings

# Codebase map (US14) — pins per plan.md; grammar smoke tests: tests/unit/test_tree_sitter_grammars.py (T101)
tree-sitter==0.23.0
tree-sitter-python==0.23.2
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
